git subrepo pull --branch=develop deps/AMICI
subrepo:
  subdir:   "deps/AMICI"
  merged:   "9e7f2ff1"
upstream:
  origin:   "git@github.com:ICB-DCM/AMICI.git"
  branch:   "develop"
  commit:   "9e7f2ff1"
git-subrepo:
  version:  "0.4.0"
  origin:   "https://github.com/ingydotnet/git-subrepo"
  commit:   "5d6aba9"
