+++ deps/AMICI/python/sdist/amici/petab_import.py
@@ -1,0 +1,1 @@
+../../amici/petab_import.py
