+++ deps/AMICI/src/CMakeLists.template.cmake
@@ -35,6 +35,13 @@ ${MODEL_DIR}/wrapfunctions.cpp
 add_library(${PROJECT_NAME} ${SRC_LIST_LIB})
 add_library(model ALIAS ${PROJECT_NAME})
 
+# This option can be helpful when using the Intel compiler and compilation of
+# wrapfunctions.cpp fails due to insufficient memory.
+option(ENABLE_WRAPFUNCTIONS_OPTIMIZATIONS "Enable compiler optimizations for wrapfunctions.cpp?" ON)
+if(NOT ENABLE_WRAPFUNCTIONS_OPTIMIZATIONS)
+    set_source_files_properties(wrapfunctions.cpp PROPERTIES COMPILE_FLAGS -O0)
+endif()
+
 target_include_directories(${PROJECT_NAME} PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}")
 
 target_link_libraries(${PROJECT_NAME}
