+++ deps/AMICI/src/interface_matlab.cpp
@@ -312,7 +312,7 @@ void setSolverOptions(const mxArray *prhs[], int nrhs, Solver &solver)
         }
 
         if (mxGetProperty(prhs[RHS_OPTIONS], 0, "newton_preeq")) {
-            solver.setNewtonPreequilibration(dbl2int(mxGetScalar(mxGetProperty(prhs[RHS_OPTIONS], 0, "newton_preeq"))));
+            solver.setPreequilibration(dbl2int(mxGetScalar(mxGetProperty(prhs[RHS_OPTIONS], 0, "newton_preeq"))));
         }
 
         if (mxGetProperty(prhs[RHS_OPTIONS], 0, "newton_maxsteps")) {
