+++ deps/AMICI/src/misc.cpp
@@ -18,7 +18,7 @@
 #endif
 
 namespace amici {
-    
+
 gsl::span<realtype> slice(std::vector<realtype> &data, const int index,
                           const unsigned size) {
     if ((index + 1) * size > data.size())
@@ -150,6 +150,39 @@ std::string backtraceString(const int maxFrames)
         trace_buf << "[truncated]\n";
 #endif
     return trace_buf.str();
+}
+
+std::string regexErrorToString(std::regex_constants::error_type err_type)
+{
+    switch (err_type) {
+    case std::regex_constants::error_collate:
+        return "error_collate";
+    case std::regex_constants::error_ctype:
+        return "error_ctype";
+    case std::regex_constants::error_escape:
+        return "error_escape";
+    case std::regex_constants::error_backref:
+        return "error_backref";
+    case std::regex_constants::error_brack:
+        return "error_brack";
+    case std::regex_constants::error_paren:
+        return "error_paren";
+    case std::regex_constants::error_brace:
+        return "error_brace";
+    case std::regex_constants::error_badbrace:
+        return "error_badbrace";
+    case std::regex_constants::error_range:
+        return "error_range";
+    case std::regex_constants::error_space:
+        return "error_space";
+    case std::regex_constants::error_badrepeat:
+        return "error_badrepeat";
+    case std::regex_constants::error_complexity:
+        return "error_complexity";
+    case std::regex_constants::error_stack:
+        return "error_stack";
+    }
+    return "unknown error";
 }
 
 } // namespace amici
