+++ deps/AMICI/tests/testSBMLSuite.py
@@ -9,14 +9,11 @@ Usage:
         test cases or 1-1780 to run all.
 """
 
-import re
 import os
 import sys
 import importlib
-import unittest
+import pytest
 import copy
-import shutil
-from typing import List
 
 import amici
 import numpy as np
@@ -25,113 +22,134 @@ import pandas as pd
 
 
 # directory with sbml semantic test cases
-test_path = os.path.join(os.path.dirname(__file__), 'sbml-test-suite', 'cases',
+TEST_PATH = os.path.join(os.path.dirname(__file__), 'sbml-test-suite', 'cases',
                          'semantic')
-upload_result_path = os.path.join(os.path.dirname(__file__),
-                                  'amici-semantic-results')
-
-ALL_TESTS = set(range(1, 1781))
 
 
-class TestAmiciSBMLTestSuite(unittest.TestCase):
-    SBML_TEST_IDS = ALL_TESTS
+@pytest.fixture(scope="session")
+def result_path():
+    # ensure directory for test results is empty
+    upload_result_path = os.path.join(os.path.dirname(__file__),
+                                      'amici-semantic-results')
+    return upload_result_path
 
-    def setUp(self):
-        self.resetdir = os.getcwd()
-        self.default_path = copy.copy(sys.path)
 
-    def tearDown(self):
-        os.chdir(self.resetdir)
-        sys.path = self.default_path
+@pytest.fixture(scope="function", autouse=True)
+def sbml_test_dir():
+    # setup
+    old_cwd = os.getcwd()
+    old_path = copy.copy(sys.path)
 
-    def runTest(self):
-        self.test_sbml_testsuite()
+    yield
 
-    def test_sbml_testsuite(self):
-        for testId in self.SBML_TEST_IDS:
-            if testId != 1395:  # we skip this test due to NaNs in the
-                # jacobian
-                with self.subTest(testId=testId):
-                    self.run_sbml_testsuite_case(get_test_str(testId))
+    # teardown
+    os.chdir(old_cwd)
+    sys.path = old_path
 
-    def run_sbml_testsuite_case(self, test_id):
-        try:
-            current_test_path = os.path.join(test_path, test_id)
 
-            # results
-            results_file = os.path.join(current_test_path,
-                                        test_id + '-results.csv')
-            results = np.genfromtxt(results_file, delimiter=',')
+def test_sbml_testsuite_case(test_number, result_path):
 
-            model, solver, wrapper = compile_model(current_test_path, test_id)
-            atol, rtol = apply_settings(current_test_path, test_id, solver,
-                                        model)
+    test_id = format_test_id(test_number)
 
-            rdata = amici.runAmiciSimulation(model, solver)
+    try:
+        current_test_path = os.path.join(TEST_PATH, test_id)
 
-            amount_species, variables_species = get_amount_and_variables(
-                current_test_path, test_id
-            )
+        # parse expected results
+        results_file = os.path.join(current_test_path,
+                                    test_id + '-results.csv')
+        results = np.genfromtxt(results_file, delimiter=',')
 
-            simulated_x = rdata['x']
-            test_x = results[1:, [
-                                     1 + wrapper.speciesIndex[variable]
-                                     for variable in variables_species
-                                     if variable in wrapper.speciesIndex.keys()
-                                 ]]
+        # setup model
+        model, solver, wrapper = compile_model(current_test_path, test_id)
+        settings = read_settings_file(current_test_path, test_id)
 
-            for species in amount_species:
-                if not species == '':
-                    symvolume = wrapper.speciesCompartment[
-                        wrapper.speciesIndex[species]
-                    ]
-                    volume = symvolume.subs({
-                        comp: vol
-                        for comp, vol in zip(
-                            wrapper.compartmentSymbols,
-                            wrapper.compartmentVolume
-                        )
-                    })
-                    volume = volume.subs({
-                        sp.Symbol(name): value
-                        for name, value in zip(
-                            model.getParameterIds(),
-                            model.getParameters()
-                        )
-                    })
+        atol, rtol = apply_settings(settings, solver, model)
 
-                    # required for 525-527, 530 as k is renamed to amici_k
-                    volume = volume.subs({
-                        sp.Symbol(name): value
-                        for name, value in zip(
-                        model.getParameterNames(),
-                        model.getParameters()
-                    )
-                    })
+        # simulate model
+        rdata = amici.runAmiciSimulation(model, solver)
 
-                    simulated_x[:, wrapper.speciesIndex[species]] = \
-                        simulated_x[:, wrapper.speciesIndex[species]] * volume
+        # verify
+        simulated_x = verify_results(settings, rdata, results, wrapper,
+                                     model, atol, rtol)
 
-            self.assertTrue(
-                np.isclose(simulated_x, test_x, atol, rtol).all()
-            )
-            print(f'TestCase {test_id} passed.')
+        print(f'TestCase {test_id} passed.')
 
-            write_result_file(simulated_x, model, test_id)
+        # record results
+        write_result_file(simulated_x, model, test_id, result_path)
 
-        except amici.sbml_import.SBMLException as err:
-            print(f'TestCase {test_id} was skipped: {err}')
+    except amici.sbml_import.SBMLException as err:
+        print(f'TestCase {test_id} was skipped: {err}')
+
+
+def verify_results(settings, rdata, results, wrapper,
+                   model, atol, rtol):
+    """Verify test results"""
+    amount_species, variables_species = get_amount_and_variables(settings)
+
+    # verify states
+    simulated_x = rdata['x']
+    expected_x = results[1:, [
+                             1 + wrapper.speciesIndex[variable]
+                             for variable in variables_species
+                             if variable in wrapper.speciesIndex.keys()
+                         ]]
+
+    concentrations_to_amounts(amount_species, wrapper, model, simulated_x)
+
+    assert np.isclose(simulated_x, expected_x, atol, rtol).all()
+
+    # TODO: verify compartment volumes and parameters
+
+    return simulated_x
+
+
+def concentrations_to_amounts(amount_species, wrapper, model, simulated_x):
+    """Convert AMICI simulated concentrations to amounts"""
+    for species in amount_species:
+        if not species == '':
+            symvolume = wrapper.speciesCompartment[
+                wrapper.speciesIndex[species]
+            ]
+            volume = symvolume.subs({
+                comp: vol
+                for comp, vol in zip(
+                    wrapper.compartmentSymbols,
+                    wrapper.compartmentVolume
+                )
+            })
+            volume = volume.subs({
+                sp.Symbol(name): value
+                for name, value in zip(
+                    model.getParameterIds(),
+                    model.getParameters()
+                )
+            })
+
+            # required for 525-527, 530 as k is renamed to amici_k
+            volume = volume.subs({
+                sp.Symbol(name): value
+                for name, value in zip(
+                    model.getParameterNames(),
+                    model.getParameters()
+                )
+            })
+
+            simulated_x[:, wrapper.speciesIndex[species]] = \
+                simulated_x[:, wrapper.speciesIndex[species]] * volume
+
 
 def write_result_file(simulated_x: np.array,
                       model: amici.Model,
-                      test_id: str):
+                      test_id: str, result_path: str):
     """
     Create test result file for upload to
     http://sbml.org/Facilities/Database/Submission/Create
 
     Requires csv file with test ID in name and content of [time, Species, ...]
     """
-    filename = os.path.join(upload_result_path, f'{test_id}.csv')
+    # TODO: only states are reported here, not compartments or parameters
+
+    filename = os.path.join(result_path, f'{test_id}.csv')
 
     df = pd.DataFrame(simulated_x)
     df.columns = model.getStateIds()
@@ -139,13 +157,16 @@ def write_result_file(simulated_x: np.array,
     df.to_csv(filename, index=False)
 
 
-def get_amount_and_variables(current_test_path, test_id):
-    settings = read_settings_file(current_test_path, test_id)
+def get_amount_and_variables(settings):
+    """Read amount and species from settings file"""
 
+    # species for which results are expected as amounts
     amount_species = settings['amount'] \
         .replace(' ', '') \
         .replace('\n', '') \
         .split(',')
+
+    # IDs of all variables for which results are expected/provided
     variables_species = settings['variables'] \
         .replace(' ', '') \
         .replace('\n', '') \
@@ -154,8 +175,8 @@ def get_amount_and_variables(current_test_path, test_id):
     return amount_species, variables_species
 
 
-def apply_settings(current_test_path, test_id, solver, model):
-    settings = read_settings_file(current_test_path, test_id)
+def apply_settings(settings, solver, model):
+    """Apply model and solver settings as specified in the test case"""
 
     ts = np.linspace(float(settings['start']),
                      float(settings['start'])
@@ -173,6 +194,7 @@ def apply_settings(current_test_path, test_id, solver, model):
 
 
 def compile_model(path, test_id):
+    """Import the given test model to AMICI"""
     sbml_file = find_model_file(path, test_id)
 
     wrapper = amici.SbmlImporter(sbml_file)
@@ -195,22 +217,23 @@ def compile_model(path, test_id):
     return model, solver, wrapper
 
 
-def find_model_file(current_test_path, testId):
+def find_model_file(current_test_path: str, test_id: str):
     """Find model file for the given test (guess filename extension)"""
-    sbmlFile = os.path.join(current_test_path, testId + '-sbml-l3v2.xml')
+
+    sbml_file = os.path.join(current_test_path, test_id + '-sbml-l3v2.xml')
 
     # fallback l3v1
-    if not os.path.isfile(sbmlFile):
-        sbmlFile = os.path.join(current_test_path, testId + '-sbml-l3v1.xml')
+    if not os.path.isfile(sbml_file):
+        sbml_file = os.path.join(current_test_path, test_id + '-sbml-l3v1.xml')
 
     # fallback l2v5
-    if not os.path.isfile(sbmlFile):
-        sbmlFile = os.path.join(current_test_path, testId + '-sbml-l2v5.xml')
+    if not os.path.isfile(sbml_file):
+        sbml_file = os.path.join(current_test_path, test_id + '-sbml-l2v5.xml')
 
-    return sbmlFile
+    return sbml_file
 
 
-def read_settings_file(current_test_path, test_id):
+def read_settings_file(current_test_path: str, test_id: str):
     """Read settings for the given test"""
     settings_file = os.path.join(current_test_path, test_id + '-settings.txt')
     settings = {}
@@ -222,42 +245,9 @@ def read_settings_file(current_test_path, test_id):
     return settings
 
 
-def get_test_str(test_id):
+def format_test_id(test_id) -> str:
+    """Format numeric to 0-padded string"""
     test_str = str(test_id)
     test_str = '0'*(5-len(test_str)) + test_str
     return test_str
 
-
-def parse_selection(selection_str: str) -> List[int]:
-    """
-    Parse comma-separated list of integer ranges, return selected indices as
-    integer list
-
-    Valid input e.g.: 1 1,3 -3,4,6-7
-    """
-    indices = []
-    for group in selection_str.split(','):
-        if not re.match(r'^(?:-?\d+)|(?:\d+(?:-\d+))$', group):
-            print("Invalid selection", group)
-            sys.exit()
-        spl = group.split('-')
-        if len(spl) == 1:
-            indices.append(int(spl[0]))
-        elif len(spl) == 2:
-            begin = int(spl[0]) if spl[0] else 0
-            end = int(spl[1])
-            indices.extend(range(begin, end + 1))
-    return indices
-
-
-if __name__ == '__main__':
-    # ensure directory for test results is empty
-    if os.path.exists(upload_result_path):
-        shutil.rmtree(upload_result_path)
-    os.mkdir(upload_result_path)
-
-    TestAmiciSBMLTestSuite.SBML_TEST_IDS = set(parse_selection(sys.argv.pop()))
-
-    suite = unittest.TestSuite()
-    suite.addTest(TestAmiciSBMLTestSuite())
-    unittest.main()
